Fix the bug of the function when index != 0
The function was supposed to traverse the array and then find the index for which the value of the array is non zero. Therefore it had to start from index and if necessary wrap back to 0 and continue until it reaches the original index value prior to the start of the traversal of the array. However the code started from ++index which is a mistake. So I rectified it.