Remove Sprite image from CardToReveal
I thought that by changing this field when the clones are made
in the scene, we would get different sprites shown to us. But I was
wrong. The Sprite component of each clone needs to be directly set.
Therefore, I removed this field and all of its instances in the code.