Fix the bug in findNonZeroElementInIntArray()
When index wraps around 3 to 0 in the second if case, it
traverses the array all the way until it reaches dummy_var.
The while loop ends without checking the case when index == dummy_var.
So that causes this function to return -1 for some of the cases.
Therefore I had to include an if statement at the end of the while
loop, in order to also check that case.