Define a public function to activate cardBack
I wanted to activate all the back cards in CardInitializer when the
player has clicked on 2 dissimilar cards. However that would not be
be possible by just having access to an instance of the cloned
CardToReveal's. Therefore I chose to have a public function that
automatically enables these game objects.